feat(mobile): replace floating nav button with fixed bottom tab bar
Swaps the draggable floating navigation button for a fixed bottom tab
bar mirroring the native app (Home, Videos, Create, Leaderboard, Menu).
Secondary routes and the login flow live in the Menu drawer, preserving
all previous navigation. Adds iOS safe-area padding.
